requirements.txt
astroid==2.4.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
